chore(coin): move shared functions and files into their own package
