+++ libs/coin-framework/package.json
@@ -56,6 +56,9 @@
     "@ledgerhq/devices": "workspace:^",
     "@ledgerhq/errors": "workspace:^",
     "@ledgerhq/hw-transport": "workspace:^",
+    "@ledgerhq/live-common-date": "workspace:^",
+    "@ledgerhq/live-common-env": "workspace:^",
+    "@ledgerhq/live-common-promise": "workspace:^",
     "@ledgerhq/logs": "workspace:^",
     "@ledgerhq/types-cryptoassets": "workspace:^",
     "@ledgerhq/types-live": "workspace:^",
@@ -69,8 +72,7 @@
     "numeral": "^2.0.6",
     "prando": "^6.0.1",
     "rxjs": "^6.6.7",
-    "rxjs-compat": "^6.6.7",
-    "utility-types": "^3.10.0"
+    "rxjs-compat": "^6.6.7"
   },
   "devDependencies": {
     "@types/invariant": "^2.2.2",
