+++ libs/coin-framework/src/network.ts
@@ -4,3 +4,13 @@ export type NetworkRequestCall = (options: {
   url?: string;
   data?: any;
 }) => Promise<any>;
+
+export const mockNetworkRequestCall =
+  (response: any): NetworkRequestCall =>
+  async (_options: {
+    method: MethodType;
+    url?: string;
+    data?: any;
+  }): Promise<any> => {
+    return Promise.resolve(response);
+  };
