+++ libs/ledger-live-common/package.json
@@ -151,6 +151,9 @@
     "@ledgerhq/hw-transport-mocker": "workspace:^",
     "@ledgerhq/hw-transport-node-speculos": "workspace:^",
     "@ledgerhq/live-app-sdk": "^0.8.1",
+    "@ledgerhq/live-common-date": "workspace:^",
+    "@ledgerhq/live-common-env": "workspace:^",
+    "@ledgerhq/live-common-promise": "workspace:^",
     "@ledgerhq/logs": "workspace:^",
     "@ledgerhq/wallet-api-core": "^0.11.0",
     "@ledgerhq/wallet-api-server": "^0.11.0",
