+++ libs/ledger-live-common/src/hw/flush.ts
@@ -1,6 +1,6 @@
 import { of } from "rxjs";
 import { withDevice } from "./deviceAccess";
-import { delay } from "@ledgerhq/coin-framework/promise";
+import { delay } from "../promise";
 
 const flush = (deviceId: string): Promise<any> =>
   delay(1000).then(() => withDevice(deviceId)(() => of(null)).toPromise());
