+++ libs/ledger-live-common/src/promise.ts
@@ -1,1 +1,1 @@
-export * from "@ledgerhq/coin-framework/promise";
+export * from "@ledgerhq/live-common-promise/index";
