+++ libs/live-common/date/package.json
@@ -1,13 +1,9 @@
 {
-  "name": "@ledgerhq/coin-polkadot",
-  "version": "0.0.2",
-  "description": "Ledger Polkadot Coin integration",
+  "name": "@ledgerhq/live-common-date",
+  "version": "0.0.1",
+  "description": "Ledger Live Date utilities",
   "keywords": [
-    "Ledger",
-    "LedgerWallet",
-    "dot",
-    "Polkadot",
-    "Hardware Wallet"
+    "Ledger"
   ],
   "repository": {
     "type": "git",
@@ -16,10 +12,13 @@
   "bugs": {
     "url": "https://github.com/LedgerHQ/ledger-live/issues"
   },
-  "homepage": "https://github.com/LedgerHQ/ledger-live/tree/develop/libs/coin-polkadot",
+  "homepage": "https://github.com/LedgerHQ/ledger-live/tree/develop/libs/ledger-live-date",
   "publishConfig": {
     "access": "public"
   },
+  "main": "lib/index.js",
+  "module": "lib-es/index.js",
+  "types": "lib/index.d.ts",
   "typesVersions": {
     "*": {
       "lib/*": [
@@ -44,27 +43,7 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@ledgerhq/coin-framework": "workspace:^",
-    "@ledgerhq/cryptoassets": "workspace:^",
-    "@ledgerhq/devices": "workspace:^",
-    "@ledgerhq/errors": "workspace:^",
-    "@ledgerhq/hw-app-polkadot": "workspace:^",
-    "@ledgerhq/live-app-sdk": "^0.8.1",
-    "@ledgerhq/live-common-date": "workspace:^",
-    "@ledgerhq/live-common-env": "workspace:^",
-    "@ledgerhq/live-common-promise": "workspace:^",
-    "@ledgerhq/logs": "workspace:^",
-    "@ledgerhq/types-cryptoassets": "workspace:^",
-    "@ledgerhq/types-live": "workspace:^",
-    "@polkadot/types": "9.8.1",
-    "@polkadot/types-known": "9.8.1",
-    "@polkadot/util": "10.3.1",
-    "@polkadot/util-crypto": "10.3.1",
-    "bignumber.js": "^9.1.0",
-    "expect": "^27.4.6",
-    "invariant": "^2.2.2",
-    "lodash": "^4.17.21",
-    "rxjs": "^6.6.7"
+    "@ledgerhq/types-live": "workspace:^"
   },
   "devDependencies": {
     "@types/jest": "^29.2.4",
@@ -86,7 +65,6 @@
     "build": "tsc && tsc -m ES6 --outDir lib-es",
     "prewatch": "pnpm build",
     "watch": "tsc --watch",
-    "doc": "documentation readme src/** --section=API --pe ts --re ts --re d.ts",
     "lint": "eslint ./src --no-error-on-unmatched-pattern --ext .ts,.tsx",
     "lint:fix": "pnpm lint --fix",
     "test": "jest"
