+++ libs/live-common/promise/src/promise.ts
