+++ package.json
@@ -30,6 +30,7 @@
     "build:llm:android": "pnpm turbo android:apk:local --no-daemon --filter=live-mobile",
     "build:llm:ios": "pnpm turbo ios:local:ipa --no-daemon --filter=live-mobile",
     "build:llm:deps": "pnpm turbo build --no-daemon --filter=\"live-mobile^...\"",
+    "build:llu": "pnpm turbo build --no-daemon --filter=\"live-utils\"",
     "build:ljs": "pnpm turbo build --no-daemon --filter=\"./libs/ledgerjs/**\"",
     "build-ci:llm:ios": "pnpm turbo ios:ci:adhoc --no-daemon --filter=live-mobile",
     "build-ci:llm:android": "pnpm turbo android:apk --no-daemon --filter=live-mobile",
@@ -58,6 +59,7 @@
     "mobile": "pnpm --filter live-mobile",
     "common": "pnpm --filter live-common",
     "common-tools": "pnpm --filter live-common-tools",
+    "utils": "pnpm --filter live-utils",
     "bot:github": "pnpm --filter live-github-bot",
     "ljs:cryptoassets": "pnpm --filter cryptoassets",
     "ljs:devices": "pnpm --filter devices",
