+++ pnpm-workspace.yaml
@@ -2,6 +2,7 @@ packages:
   - "apps/*"
   - "apps/ledger-live-desktop/tests/utils/**"
   - "libs/*"
+  - "libs/live-common/*"
   - "libs/ledgerjs/**"
   - "libs/ui/packages/*"
   - "libs/ledger-live-common/tools"
